chore: move `Opts` to `cargo_near_build::types::near::build`
