+++ cargo-near-build/src/types/near/build.rs
@@ -1,82 +1,113 @@
-use crate::env_keys;
+use std::marker::PhantomData;
 
-pub type Version = String;
+use camino::Utf8PathBuf;
+use sha2::{Digest, Sha256};
 
-#[derive(Debug)]
-pub enum VersionMismatch {
-    Some {
-        environment: String,
-        current_process: String,
-    },
-    None,
-    UnknownFromDocker,
+use crate::{types::color_preference::ColorPreference, ArtifactType, WASM};
+
+pub mod version_mismatch;
+
+pub struct CompilationArtifact<T: ArtifactType = WASM> {
+    pub path: Utf8PathBuf,
+    pub fresh: bool,
+    pub from_docker: bool,
+    pub cargo_near_version_mismatch: version_mismatch::VersionMismatch,
+    pub artifact_type: PhantomData<T>,
 }
 
-impl std::fmt::Display for VersionMismatch {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Some {
-                environment,
-                current_process,
-            } => {
-                write!(
-                    f,
-                    "`cargo-near` version {} -> `cargo-near` environment version {}",
-                    current_process, environment
-                )
-            }
-            Self::None => write!(f, "no `cargo-near` version mismatch in nested builds detected",),
-            Self::UnknownFromDocker => write!(f, "it's unknown if `cargo-near` version mismatch has occurred in docker build environment",),
-        }
+impl<T: ArtifactType> CompilationArtifact<T> {
+    pub fn compute_hash(&self) -> eyre::Result<SHA256Checksum> {
+        let mut hasher = Sha256::new();
+        hasher.update(std::fs::read(&self.path)?);
+        let hash = hasher.finalize();
+        let hash: &[u8] = hash.as_ref();
+        Ok(SHA256Checksum {
+            hash: hash.to_vec(),
+        })
+    }
+}
+// TODO: make non-pub
+pub struct SHA256Checksum {
+    pub hash: Vec<u8>,
+}
+
+impl SHA256Checksum {
+    pub fn to_hex_string(&self) -> String {
+        hex::encode(&self.hash)
+    }
+
+    pub fn to_base58_string(&self) -> String {
+        bs58::encode(&self.hash).into_string()
     }
 }
 
-impl VersionMismatch {
-    fn current_version() -> Version {
-        format!("{} {}", "cargo-near-build", env!("CARGO_PKG_VERSION"))
-    }
+#[derive(Debug, Default, Clone)]
+pub struct Opts {
+    /// disable implicit `--locked` flag for all `cargo` commands, enabled by default
+    pub no_locked: bool,
+    /// Build contract in debug mode, without optimizations and bigger is size
+    pub no_release: bool,
+    /// Do not generate ABI for the contract
+    pub no_abi: bool,
+    /// Do not embed the ABI in the contract binary
+    pub no_embed_abi: bool,
+    /// Do not include rustdocs in the embedded ABI
+    pub no_doc: bool,
+    /// Copy final artifacts to this directory
+    pub out_dir: Option<camino::Utf8PathBuf>,
+    /// Path to the `Cargo.toml` of the contract to build
+    pub manifest_path: Option<camino::Utf8PathBuf>,
+    /// Set compile-time feature flags.
+    pub features: Option<String>,
+    /// Disables default feature flags.
+    pub no_default_features: bool,
+    /// Coloring: auto, always, never
+    pub color: Option<ColorPreference>,
+}
 
-    pub fn export_builder_and_near_abi_versions() {
-        if std::env::var(env_keys::CARGO_NEAR_VERSION).is_err() {
-            std::env::set_var(env_keys::CARGO_NEAR_VERSION, Self::current_version());
+impl Opts {
+    /// this is just 1-to-1 mapping of each struct's field to a cli flag
+    /// in order of fields, as specified in struct's definition.
+    /// `Default` implementation corresponds to plain `cargo near build` command without any args
+    // TODO: make pub(crate)
+    pub fn get_cli_build_command(&self) -> Vec<String> {
+        let mut cargo_args = vec!["cargo", "near", "build"];
+        if self.no_locked {
+            cargo_args.push("--no-locked");
         }
-        if std::env::var(env_keys::CARGO_NEAR_ABI_SCHEMA_VERSION).is_err() {
-            std::env::set_var(
-                env_keys::CARGO_NEAR_ABI_SCHEMA_VERSION,
-                near_abi::SCHEMA_VERSION,
-            );
+        // `no_docker` field isn't present
+        if self.no_release {
+            cargo_args.push("--no-release");
         }
-    }
-
-    pub fn get_coerced_builder_version() -> eyre::Result<(Version, VersionMismatch)> {
-        match std::env::var(env_keys::CARGO_NEAR_ABI_SCHEMA_VERSION) {
-            Ok(env_near_abi_schema_version) => {
-                if env_near_abi_schema_version != near_abi::SCHEMA_VERSION {
-                    return Err(eyre::eyre!(
-                        "current process NEAR_ABI_SCHEMA_VERSION mismatch with env value: {} vs {}",
-                        near_abi::SCHEMA_VERSION,
-                        env_near_abi_schema_version,
-                    ));
-                }
-            }
-            Err(_err) => {}
+        if self.no_abi {
+            cargo_args.push("--no-abi");
         }
-        let current_version = Self::current_version();
-
-        let result = match std::env::var(env_keys::CARGO_NEAR_VERSION) {
-            Err(_err) => (current_version.to_string(), VersionMismatch::None),
-            Ok(env_version) => match env_version == current_version {
-                true => (current_version.to_string(), VersionMismatch::None),
-                // coercing to env_version on mismatch
-                false => (
-                    env_version.clone(),
-                    VersionMismatch::Some {
-                        environment: env_version,
-                        current_process: current_version.to_string(),
-                    },
-                ),
-            },
-        };
-        Ok(result)
+        if self.no_embed_abi {
+            cargo_args.push("--no-embed-abi");
+        }
+        if self.no_doc {
+            cargo_args.push("--no-doc");
+        }
+        if let Some(ref out_dir) = self.out_dir {
+            cargo_args.extend_from_slice(&["--out-dir", out_dir.as_str()]);
+        }
+        if let Some(ref manifest_path) = self.manifest_path {
+            cargo_args.extend_from_slice(&["--manifest-path", manifest_path.as_str()]);
+        }
+        if let Some(ref features) = self.features {
+            cargo_args.extend(&["--features", features]);
+        }
+        if self.no_default_features {
+            cargo_args.push("--no-default-features");
+        }
+        let color;
+        if let Some(ref color_arg) = self.color {
+            color = color_arg.to_string();
+            cargo_args.extend(&["--color", &color]);
+        }
+        cargo_args
+            .into_iter()
+            .map(|el| el.to_string())
+            .collect::<Vec<_>>()
     }
 }
