+++ cargo-near/Cargo.toml
@@ -24,8 +24,6 @@ eula = false
 
 [dependencies]
 cargo-near-build = { version = "0.1.0", path = "../cargo-near-build" }
-camino = "1.1.1"
-cargo_metadata = "0.18.1"
 clap = { version = "4.0.18", features = ["derive", "env"] }
 colored = "2.0"
 env_logger = "0.11.5"
@@ -44,7 +42,6 @@ shell-words = "1.0.0"
 interactive-clap = "0.3"
 interactive-clap-derive = "0.3"
 near-cli-rs = { version = "0.14.0", default-features = false }
-dunce = "1"
 tempfile = "3.10.1"
 git2 = "0.19"
 home = "0.5.9"
