+++ cargo-near/src/build_extended/build_script.rs
@@ -1,7 +1,7 @@
 use std::marker::PhantomData;
 
-use crate::BuildArtifact;
-use cargo_near_build::types::near::build::VersionMismatch;
+use cargo_near_build::camino;
+use cargo_near_build::{types::near::build::version_mismatch::VersionMismatch, BuildArtifact};
 use rustc_version::Version;
 
 /// `cargo::` prefix for build script outputs, that `cargo` recognizes
