+++ cargo-near/src/commands/build_command/docker/cloned_repo.rs
@@ -1,16 +1,14 @@
 use std::marker::PhantomData;
 
 use crate::commands::build_command::ArtifactMessages;
-use camino::Utf8PathBuf;
+use cargo_near_build::camino;
 use cargo_near_build::{
     pretty_print,
-    types::{
-        cargo::{
-            manifest_path::{ManifestPath, MANIFEST_FILE_NAME},
-            metadata::CrateMetadata,
-        },
-        near::{build::VersionMismatch, CompilationArtifact},
+    types::cargo::{
+        manifest_path::{ManifestPath, MANIFEST_FILE_NAME},
+        metadata::CrateMetadata,
     },
+    BuildArtifact, VersionMismatch,
 };
 use colored::Colorize;
 
@@ -76,7 +74,7 @@ impl ClonedRepo {
     pub(super) fn copy_artifact(
         self,
         cli_override: Option<crate::types::utf8_path_buf::Utf8PathBuf>,
-    ) -> color_eyre::eyre::Result<CompilationArtifact> {
+    ) -> color_eyre::eyre::Result<BuildArtifact> {
         let tmp_out_dir = self.tmp_crate_metadata.resolve_output_dir(None)?;
 
         let destination_crate_metadata = {
@@ -96,10 +94,10 @@ impl ClonedRepo {
 }
 
 fn copy(
-    tmp_out_dir: Utf8PathBuf,
+    tmp_out_dir: camino::Utf8PathBuf,
     tmp_crate_metadata: CrateMetadata,
-    mut destination_dir: Utf8PathBuf,
-) -> color_eyre::eyre::Result<CompilationArtifact> {
+    mut destination_dir: camino::Utf8PathBuf,
+) -> color_eyre::eyre::Result<BuildArtifact> {
     println!(
         " {} {}",
         "artifact search location in temporary build site:".green(),
@@ -126,7 +124,7 @@ fn copy(
         std::fs::remove_file(&out_wasm_path)?;
     }
     std::fs::copy::<camino::Utf8PathBuf, camino::Utf8PathBuf>(in_wasm_path, out_wasm_path.clone())?;
-    let result = CompilationArtifact {
+    let result = BuildArtifact {
         path: out_wasm_path,
         fresh: true,
         from_docker: true,
