+++ cargo-near/src/commands/build_command/docker/crate_in_repo.rs
@@ -1,3 +1,4 @@
+use cargo_near_build::camino;
 use color_eyre::eyre::ContextCompat;
 use colored::Colorize;
